backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
spotipy==2.23.0
pandas==2.1.1
python-dotenv==1.0.0
requests==2.31.0
numpy==1.25.2
pyopenssl==23.3.0